% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersja_temp.R
\name{konwersja_temp}
\alias{konwersja_temp}
\title{Konwersja temperatur}
\usage{
konwersja_temp(temperatura_f)
}
\arguments{
\item{temperatura_f}{temperatura w F}
}
\value{
wektor o takiej samej dlugosci jak temperatura_f
}
\description{
Funkcja przelicza temp z F na C
}
\examples{
konwersja_temp(99)
}
