% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moj_df.R
\docType{data}
\name{moj_df}
\alias{moj_df}
\title{Example data}
\format{
\subsection{\code{moj_df}}{

A data frame with 20 observations on 2 variables.
\describe{
\item{x}{numeric vector}
\item{y}{numeric vector}
}
}
}
\source{
Manually created.
}
\usage{
moj_df
}
\description{
Example data created manually for the purpose of this package.
}
\keyword{datasets}
