% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers}
\usage{
remove_outliers(x, lq = 0.25, uq = 0.75)
}
\arguments{
\item{x}{vector of values}

\item{lq}{lower quantile}

\item{uq}{upper quantile}
}
\value{
vector
}
\description{
Function to remove outliers observation based
on a quantile distance
}
\examples{
data <- c(1,1,1,1,1,99)
new_data <- remove_outliers(data)
print(data)
print(new_data)

data("moj_df", package = "mojpakiet")
remove_outliers(moj_df$x)

system.file("moj_df.csv", package = "mojpakiet")
remove_outliers(moj_df$y)
}
